Move sendFail(Malformed)
